requirements.txt
asgiref==3.6.0
bottle==0.12.23
certifi==2022.12.7
charset-normalizer==3.0.1
cmake==3.25.2
colorama==0.4.6
conan==1.58.0
Django==4.1.4
django-crispy-forms==1.14.0
fasteners==0.18
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mcculw==1.0.0
MPC2860==0.3
ninja==1.11.1
node-semver==0.6.1
numpy==1.24.1
numpy-stl==3.0.0
patch-ng==1.17.4
pluginbase==1.0.1
protobuf==4.21.12
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-utils==3.4.5
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.14
